--- src/infrastructure/dto/strapi-common-dto.ts
@@ -1,10 +1,0 @@
-import { z } from 'zod';
-
-export const strapiCommonAttributesDTO = z.object({
-  createdAt: z.string().datetime(),
-  updatedAt: z.string().datetime(),
-});
-
-export const strapiCommonDTO = z.object({
-  id: z.number(),
-});
