+++ src/infrastructure/problem/problem-repository.ts
@@ -3,10 +3,10 @@ import { Problem } from '@/domain/entities/problem';
 import { ResultAsync, err, ok } from 'neverthrow';
 import { normalizeError } from '@/lib/err-utils';
 import { WithJson, withJson } from '../infra-utils';
-import { StrapiError } from '../strapi-error';
+import { StrapiError } from '../strapi/strapi-error';
 import { strapiProblems } from './problem-response';
-import { errorResponse } from '../dto/error';
-import { ProblemDTO } from '../dto/problem-dto';
+import { strapiErrorDTO } from '../strapi/strapi-response';
+import { ProblemDTO } from './problem-response';
 import { STRAPI_API_URL } from '@/constants/paths';
 
 export const newProblemFromDTO = (problem: ProblemDTO) => {
@@ -61,7 +61,7 @@ const getStrapiErrorFromGet = (res: WithJson<Response>): StrapiError => {
   const status = res.status;
   const body = res.js;
 
-  const parsed = errorResponse.safeParse(body);
+  const parsed = strapiErrorDTO.safeParse(body);
 
   const message = parsed.success ? parsed.data.error.message : 'Unknown error';
